// data/github/current/ui/packages/pull-request-viewer/components/activity/timeline-items/__generated__/ReviewRequestedEventAssignedFromTeamStoryQuery.graphql.ts
export type ReviewRequestedEventAssignedFromTeamStoryQuery$variables = {
  id: string;
};
export type ReviewRequestedEventAssignedFromTeamStoryQuery$data = {
  readonly reviewRequestedEvent: {
    readonly " $fragmentSpreads": FragmentRefs<"ReviewRequestedEvent_reviewRequestedEvent">;
  } | null | undefined;
};
export type ReviewRequestedEventAssignedFromTeamStoryQuery = {
  response: ReviewRequestedEventAssignedFromTeamStoryQuery$data;
  variables: ReviewRequestedEventAssignedFromTeamStoryQuery$variables;
};